Refactor: Update database interactions and UI for account management
- Changed `id` type from `Int` to `Long` in `ContaDao.getContaById` and related calls in `PaginaInicialViewModel` and `MainActivity2`.
- Introduced `LocalDate` to `String` converters in `DateConverter.kt`.
- Modified `BreezeCard` in `PaginaInicial.kt` to pass `apagarConta` lambda directly, removing `viewModel` dependency from the component.
- Updated `BreezeCard` to accept an `apagarConta` lambda.
- Changed `idContaPai` type in `ParcelaEntity` from `String` to `Long`.
- Updated `MainActivity2` to use `collectAsStateWithLifecycle` and handle nullable `conta` state.
- Passed only the `Conta` object to `EditarValorConta`, removing the `id` parameter. Initialized `text` field with current account value.
- Added `isContaParcelada` field to `Conta` object creation in `AdicionarContaViewModel.salvaContaDatabase`.
- Implemented `salvaParcelasDatabase` in `AdicionarContaViewModel` to save installment data when an account is marked as parceled.
- Incremented database version to 7 in `BreezeDatabase.kt`.

+++ app/src/main/java/com/migueldk17/breeze/dao/ContaDao.kt
@@ -15,7 +15,7 @@ interface ContaDao {
      fun getContas(): Flow<List<Conta>>
      //Terminar comando SQL e partir pro ViewModel
      @Query("SELECT * FROM conta_table WHERE id = :id LIMIT 1")
-     suspend fun getContaById(id: Int): Conta
+     suspend fun getContaById(id: Long): Conta
      //Insere a conta no Room
      @Insert(onConflict = OnConflictStrategy.REPLACE) //Caso haja conflito de IDS a mais recente subistitui a mais antiga
     suspend fun insertConta(conta: Conta): Long
